Update CMakeLists.txt: 1. Eigen dependency description 2. Cmake Policy 0046

+++ aerial_robot_base/src/kf/xy_roll_pitch_bias_plugin.cpp
@@ -84,7 +84,7 @@ namespace kf_plugin
     void updatePredictModel(const vector<double>& params)
     {
       if(params.size() != 7)
-        ROS_INFO("params.size is: %d", params.size());
+        ROS_INFO("params.size is: %d", (int)params.size());
       assert(params.size() == 7);
 
       float dt = params[0];
